수정 로직 구현

+++ src/main/java/git_kkalnane/starbucksbackenv2/domain/cart/dto/request/CartItemDto.java
@@ -17,7 +17,7 @@ public record CartItemDto(
        BeverageTemperatureOption temperatureOption ,
        @NotNull List<CartItemOptionDto> cartItemOptions,
        BeverageSizeOption cupSize,
-       @Min(1) Long quantity,
+       @Min(1) int quantity,
        Long priceWithOptions
 
 ) {
